feat(project): project detail show workflow list
