+++ web/src/components/workflow/component/list/index.js
@@ -1,0 +1,2 @@
+import ListLayout from './ListLayout';
+export default ListLayout;
